Use Value to share section code
